# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "EnvironmentName": {
        "description": "One of the two environments every project has.",
        "enum": [
          "staging",
          "production"
        ],
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable, machine-readable code, e.g. `DECISION_NOT_FOUND`.",
            "type": "string"
          },
          "details": {
            "description": "Engine error document (failing node and reason) for evaluation failures."
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "message": {
            "description": "Safe to show to a person.",
            "type": "string"
          },
          "requestId": {
            "description": "Same value as the `x-request-id` response header; quote it when reporting a problem.",
            "type": "string"
          }
        },
        "required": [
          "message",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "FeedBatch": {
        "description": "A batch of records, as a Runtime sends it (docs/decision-log-feed.md).",
        "properties": {
          "records": {
            "description": "At most 1000 records; each is checked on its own.",
            "items": {},
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "type": "object"
      },
      "FeedReceipt": {
        "properties": {
          "accepted": {
            "description": "Records stored, now or before (a record sent twice is stored once).",
            "minimum": 0,
            "type": "integer"
          },
          "rejected": {
            "description": "Records not stored, with why: `invalid`, `wrong_environment` or `unknown_release`.",
            "items": {
              "$ref": "#/components/schemas/FeedRejection"
            },
            "type": "array"
          }
        },
        "required": [
          "accepted",
          "rejected"
        ],
        "type": "object"
      },
...
      "IssueLogTokenRequest": {
        "properties": {
          "environment": {
            "$ref": "#/components/schemas/EnvironmentName"
          },
          "name": {
            "description": "1 to 100 characters, e.g. the Runtime's host.",
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "environment",
          "name"
        ],
        "type": "object"
      },
...
      "IssuedLogTokenResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LogTokenResponse"
          },
          {
            "properties": {
              "token": {
                "description": "Shown this once: set it as the Runtime's `DECISION_LOG__TOKEN`.",
                "type": "string"
              }
            },
            "required": [
              "token"
            ],
            "type": "object"
          }
        ]
      },
...
      "LogTokenListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LogTokenResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane of Donka: authoring, simulation, releases.",
    "license": {
      "name": ""
    },
    "title": "Donka Studio API",
    "version": "unknown"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/decision-log/records": {
      "post": {
        "operationId": "receive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedReceipt"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not a batch, or more than 1000 records (INVALID_REQUEST, TOO_MANY_RECORDS)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, unknown or revoked token (INVALID_TOKEN)"
          }
        },
        "summary": "Receives the records a Runtime sends (`Authorization: Bearer <decision-log token>`).",
        "tags": [
          "decision-log"
        ]
      }
    },
    "/decision-log/tokens": {
      "get": {
        "operationId": "tokens",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogTokenListResponse"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not an administrator (FORBIDDEN)"
          }
        },
        "summary": "Every decision-log token, live ones first; never their value (administrators).",
        "tags": [
          "decision-log"
        ]
      },
      "post": {
        "operationId": "issue_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueLogTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedLogTokenResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or too long name (INVALID_REQUEST)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not an administrator (FORBIDDEN)"
          }
        },
        "summary": "Issues a token Runtimes of one environment send their records with\n(administrators). The token is in this response only.",
        "tags": [
          "decision-log"
        ]
      }
    },
    "/decision-log/tokens/{token_id}": {
      "delete": {
        "operationId": "revoke_token",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not an administrator (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown or already revoked (TOKEN_NOT_FOUND)"
          }
        },
        "summary": "Revokes a decision-log token: batches sent with it are refused from now on (administrators).",
        "tags": [
          "decision-log"
        ]
      }
    },
...
    "/projects/{project_id}/decision-log": {
      "get": {
        "operationId": "search",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "The caller's reference (`X-Donka-Reference`), exactly.",
            "in": "query",
            "name": "reference",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only this decision, by key.",
            "in": "query",
            "name": "decisionKey",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only this outcome, exactly (`error` for failed evaluations).",
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EnvironmentName"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecordStatus"
            }
          },
          {
            "description": "Evaluated from this instant, inclusive (ISO-8601).",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Evaluated until this instant, exclusive (ISO-8601).",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Page size, 1 to 100 (default 50).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordListResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A filter is not valid (INVALID_REQUEST)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          }
        },
        "summary": "The project's decision records, newest first (any member). Lists hold\nnothing the decisions read or answered.",
        "tags": [
          "decision-log"
        ]
      }
    },
    "/projects/{project_id}/decision-log/settings": {
      "get": {
        "operationId": "settings",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionLogSettingsBody"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          }
        },
        "summary": "The project's decision-log settings (any member).",
        "tags": [
          "decision-log"
        ]
      },
      "put": {
        "operationId": "update_settings",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionLogSettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionLogSettingsBody"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not a dotted path of field names (INVALID_REQUEST)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Only owners change settings (FORBIDDEN)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_ARCHIVED"
          }
        },
        "summary": "Names the output field read as each new record's outcome (owners).\nRecords already stored keep the outcome they arrived with.",
        "tags": [
          "decision-log"
        ]
      }
    },
    "/projects/{project_id}/decision-log/{record_id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or RECORD_NOT_FOUND"
          }
        },
        "summary": "Opens a decision record (any member). Each opening is audited\n(`decision_record.viewed`): records hold applicants' personal data.",
        "tags": [
          "decision-log"
        ]
      }
    },
    "/projects/{project_id}/decision-log/{record_id}/replay": {
      "post": {
        "operationId": "replay",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "PROJECT_NOT_FOUND or RECORD_NOT_FOUND"
          }
        },
        "summary": "Evaluates a record again with the release that answered it (any member);\nconnector nodes answer with what they answered then, so nothing is called.\nAudited (`decision_record.replayed`).",
        "tags": [
          "decision-log"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}